requirements.txt
requests>=2.28.0
python-dateutil>=2.8.0
pyyaml>=6.0
